# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Integrity-First - OpenAPI 3.0",
    "description": "This is a Intergity-First Server based on the OpenAPI 3.0 specification.",
    "contact": {
      "email": "taha.muhammad.official@gmail.com"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:3002"
    },
    {
      "url": "https://staging.integrity1auto.com"
    },
    {
      "url": "https://api.integrity1auto.com"
    }
  ],
...
  "paths": {
...
    "/users/post-register": {
      "post": {
        "tags": ["Register"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Call this API after phone verification to set password and use referral code",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [],
                "properties": {
                  "fullName": {
                    "type": "string",
                    "example": "Taha Muhammad"
                  },
                  "referralCode": {
                    "type": "string",
                    "example": "JHJKASHDKSHA"
                  },
                  "email": {
                    "type": "string",
                    "example": "taha.muhammad.official@gmail.com"
                  },
                  "deviceType": {
                    "type": "string",
                    "example": "IOS | ANDROID"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/admin/users": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Admin"],
        "summary": "Create user",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phoneNumber",
                  "email",
                  "password",
                  "fullName",
                  "shopId"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "taha@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "12sdkljfisd78979cx8v"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+923204858850"
                  },
                  "fullName": {
                    "type": "string",
                    "example": "Taha Muhammad"
                  },
                  "shopId": {
                    "type": "string",
                    "example": "63e0b5a0e22a079f23e9beb0"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 10
            }
          }
        ],
        "tags": ["Admin"],
        "summary": "Get users list",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/admin/users/{userId}": {
      "patch": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": ["Admin"],
        "summary": "Update user",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/users": {
      "patch": {
        "tags": ["Profile"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Call this API to update user's details",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "tags": ["Profile"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Call this API to delete 1 or more users",
        "description": "",
        "parameters": [
          {
            "name": "userIds",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/users/me": {
      "get": {
        "tags": ["Profile"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Call this API to get logged-in user details",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/users/{userId}/block": {
      "post": {
        "tags": ["Admin"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Call this API to block a user",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Call this API to unblock a user",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/users/token": {
      "delete": {
        "tags": ["Firebase Token"],
        "summary": "Remove user firebase token",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "deviceId",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "12897897987912738ashjdgjhsagdas678d67a8sd"
            }
          }
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/users/notifications/read-all": {
      "post": {
        "tags": ["Notification"],
        "summary": "Read all notifications of current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/users/notifications": {
      "get": {
        "tags": ["Notification"],
        "summary": "Get the list of paginated notifications of current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 10
            }
          }
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/users/reward-transaction-history": {
      "get": {
        "tags": ["Reward Transaction History"],
        "summary": "Get the list of reward transaction history of current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "WebhookAPIKeyAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}